+++ src/components/modal/Modal.module.css
@@ -23,13 +23,15 @@ p {
   display: flex;
   flex-direction: column;
   gap: 16px;
-  z-index: 1000;
+  z-index: 999;
   position: fixed;
   top: 0;
   right: 0;
+  min-width: 400px;
   width: 40vw;
   height: 100vh;
   background-color: var(--global-baby-blue);
+  overflow-y: auto;
 }
 
 .modal div {
@@ -44,14 +46,19 @@ p {
   border-bottom: solid 1px var(--global-black);
 }
 
-.modal .modalHeader b {
+.modal .modalHeader .iconFechar {
+  height: 25px;
   width: auto;
   cursor: pointer;
-  padding: 0px;
 }
 
 .line {
   padding: 0px 16px;
+}
+
+.modal .iconDoacoes {
+  height: 100%;
+  width: 100%;
 }
 
 .background {
@@ -78,7 +85,7 @@ p {
   padding: 0px;
   justify-content: center;
   border-bottom: 1px solid var(--global-light-grey);
-} 
+}
 
 .titulo h1 {
   padding: 0px;
@@ -97,7 +104,7 @@ p {
 }
 
 .botao {
-  width: fit-content; 
+  width: fit-content;
   align-self: flex-end;
   height: 100%;
   border-style: none;
@@ -109,3 +116,32 @@ p {
   background-color: var(--global-blue);
 }
 
+.informacoesDoacao {
+  height: 100%;
+  display: flex;
+  align-items: center;
+  padding: 16px;
+  gap: 16px;
+
+}
+
+.modalImagem {
+  width: 100%;
+  height: auto;
+  max-width: 5vw;
+}
+
+.modalImagem img {
+  margin: auto;
+  max-width: 100%;
+  height: auto;
+  object-fit: cover;
+}
+
+
+@media (max-width: 400px) {
+  .modal {
+    min-width: auto;
+    width: 100vw;
+  }
+}
